+++ software/contrib/menu.py
@@ -3,6 +3,7 @@ from bootloader import BootloaderMenu
 
 from contrib.coin_toss import CoinToss
 from contrib.consequencer import Consequencer
+from contrib.cvecorder import CVecorder
 from contrib.diagnostic import Diagnostic
 from contrib.hamlet import Hamlet
 from contrib.harmonic_lfos import HarmonicLFOs
@@ -18,6 +19,7 @@ from calibrate import Calibrate
 EUROPI_SCRIPT_CLASSES = [
     CoinToss,
     Consequencer,
+    CVecorder,
     Diagnostic,
     Hamlet,
     HarmonicLFOs,